:recycle: Refactor current structure to allow clean set up for corona overview
