+++ frontend/src/pages/CoronaOverview.js
@@ -1,0 +1,11 @@
+import styled from 'styled-components/macro'
+import PageHeader from "../components/PageLayoutComponents/PageHeader";
+import OverviewPageLayout from "../components/PageLayoutComponents/OverviewPageLayout";
+
+export default function CoronaOverview() {
+    return (
+        <OverviewPageLayout>
+            <PageHeader>Corona Lage</PageHeader>
+        </OverviewPageLayout>
+    )
+}
